RHICOMPL-230 - Empty state on main page shouldn't show tabs
